Add support for frame_size unequal 1024

+++ src/wave.rs
@@ -1,7 +1,7 @@
 use crate::ctrl_msg::CtrlMsg;
 
 pub trait Wave: Send {
-    fn new(frame_size: u32, sample_rate: f32) -> Self
+    fn new(frame_size: usize, sample_rate: f32) -> Self
     where
         Self: Sized;
     fn process_samples(&mut self, _output_l: &mut [f32], _output_r: &mut [f32]) {}
